fix:int64

+++ pkg/apiObject/pod.go
@@ -65,8 +65,8 @@ type ContainerLifecycle struct {
 // 关于CPU和Memory怎么写，看这里
 // https://kubernetes.io/zh-cn/docs/concepts/configuration/manage-resources-containers/
 type ContainerResourcesTypes struct {
-	CPU    int64 `yaml:"cpu" json:"cpu"`       // 代表CPU的占比，最大是10^9
-	Memory int64 `yaml:"memory" json:"memory"` // 代表内存的占比，单位是byte
+	CPU    int `yaml:"cpu" json:"cpu"`       // 代表CPU的占比，最大是10^9
+	Memory int `yaml:"memory" json:"memory"` // 代表内存的占比，单位是byte
 }
 
 // 这个当你为 Pod 中的 Container 指定了资源 请求时， kube-scheduler 就利用该信息决定将 Pod 调度到哪个节点上。
